constrain minimum train set length for SU jobs
Breaking the train set apart during a SU job when one or more of the resulting train sets will be shorter than the minimum length required for jobs results in unnecessary work that must simply be undone by the next SL job.
This feels pointless (because it is), and that's NO FUN.
We can prevent this unnecessary work by constraining the minimum number of cars that a SU job may place on any one track to the minimum number of cars required for a job.