Show floaty messages for critical failures (#26)
This change introduces showing in-game messages to the player when a critical failure has occurred and the game will be closed. This is an improvement over only logging the failure to the log file as the player will have more context to understand why the game has quit, which could otherwise appear as if it were a crash.

Based on and code from [dv-handbrake](https://github.com/mspielberg/dv-handbrake/blob/03a5d3ef1206b297cbf8dc9dc4739852e5ec7acf/HandBrake.cs#L79-L115).