Clear both registry and existing car IDs
Directly calling clear on the registry List didn't remove the IDs from IdGenerator causing ID collisions when the next save data was loaded.
Using RollingStockManager's Remove method instead properly removes both the equipment record and the ID entry.
SingletonBehaviour was causing the rolling stock registry List to be recreated everytime a save was loaded, making it not possible to properly clear all the necessary entries (in IdGenerator).
Switching to a more traditional singleton pattern instead preserves the records in the rolling stock registry so they can be properly removed when the next save data is loaded.