Still clear old data even if there's nothing new to load
Leaving the handling of clearing old data to the load data functions opens the door to unexpected behavior if a save without RSO data is loaded after a save with RSO data.
Moving the handling of data clearing to SaveManager allows the old data to be cleared even if there's no incoming new data.