Make LH jobs pay nothing
This is done so LH jobs don't have a monetary influence on the player's decisions.
They are intended to serve only as a suggestion of where the associated wagons would be able to receive a SL job.
This is a game mechanic aimed at helping newer players discover where the different wagon types that they buy can be used.
RSO still prefers generating a SL job over generating a LH job.