Add logistic hauls (#60)
Some players have reported not knowing where to take wagons to get jobs for them. Other players have reported job generation as not working, but the station they were in didn't support the wagons they had. This aims to alleviate these issues by giving suggestions for where players may take wagons to get jobs generating for them.

The LH jobs have had their payouts reduced to $0 so that they act merely as suggestions. Payouts would give an incentive to deliver the empty wagons to a particular station, but players should feel no pressure to take the wagons where the LH job says to. I.e. it should not feel like a penalty to choose to take wagons to a different station.

* Find wagons for logistic jobs

* Generate logistic haul jobs

* Make LH jobs pay nothing

This is done so LH jobs don't have a monetary influence on the player's decisions.
They are intended to serve only as a suggestion of where the associated wagons would be able to receive a SL job.
This is a game mechanic aimed at helping newer players discover where the different wagon types that they buy can be used.
RSO still prefers generating a SL job over generating a LH job.