Move job generation to base chain controller class
Now that LH jobs are present, it's necessary to apply the patch that handles job generation to the job chain controller base class.
This is because the base game uses the standard job chain controller class for LH jobs.

This commit also changes RSO LH jobs to use the standard job chain controller to be in alignment with the base game.