Prompt player to choose starter equipment
Having the player configure a mod setting before first loading a save with the mod active to choose a starting locomotive was obtuse.
Prompting the player using the same popup messages that were being displayed anyway also provides an opportunity to give the player some wagons to start with rather than relying on the player to buy wagons using a cash starting bonus.