Fix logger arguments (#65)
Some logger methods that excepted a string were being passed a message factory instead. This fixes those calls and updates the logger methods to prevent such accidents in the future.

This also sets the default log level for release builds to Info so that logs contain sufficient information to investigate reported issues.