Ensure starter equipment is movable
There was a weird bug reported by Chris on the Altfuture discord server wherein the starter wagons would spawn in an immovable state.
This fixes that by forcing the starter equipment to respawn.
I'm not sure why the wagons were becoming immovable or why this works to fix that, but it does. ¯\_(ツ)_/¯